requirements.txt
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
PyYAML==6.0
